pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "personal-advocacy-toolkit"
version = "0.0.0"
description = "Evidence-integrity and packet-assembly tooling for non-technical people organizing fact-heavy disputes."
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
license-files = ["LICENSE"]
authors = [{ name = "advocacy-toolkit contributors" }]
keywords = [
    "legal-tech",
    "civic-tech",
    "evidence",
    "chain-of-custody",
    "consumer-protection",
]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Legal Industry",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    # Core runtime. Kept small on purpose; most evidence-integrity tools
    # (1A: evidence_hash, provenance, provenance_snapshot, hooks) use stdlib
    # only. Below are the deps introduced by Phase 1 packet + manifest tools.
    "PyYAML>=6",          # packet-manifest + correspondence-manifest config loading
    "reportlab>=4",       # packet cover/separator/watermark PDF generation
    "pypdf>=4",           # PDF merge + read for packet assembly
    "pikepdf>=8",         # deterministic PDF metadata + /ID control (compile_reference)
    "Jinja2>=3",          # scripts/letters/draft.py — template rendering
    "python-docx>=1",     # scripts/letters/draft.py — .docx output
    "defusedxml>=0.7",    # safe XML parsing for docx_catalog / docx_apply_replies
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "ruff>=0.4",
]
synthetic-case = [
    # Needed to regenerate the Maryland-Mustang synthetic artifacts
    # (valuation PDF, photo placeholders, MIA complaint .docx) that Phase 2
    # currently ships as .md fallbacks.
    "Pillow>=10",
    "python-docx>=1",
]
publish = [
    # Phase 3C publication-safety tools.
    # - Pillow: exif_scrub.py (also used by the synthetic-case extra)
    # - reportlab: used by tests to synthesize PDF fixtures for pdf_redact
    # - git-filter-repo is a separate binary; install via your package
    #   manager (`brew install git-filter-repo`, etc.). history_sanitizer.py
    #   shells out to it.
    "Pillow>=10",
    "PyYAML>=6",
    "pypdf>=4",
    "reportlab>=4",
]
llm = [
    # Optional: enables `python -m scripts.case_map_build --llm` to enrich
    # dashboard synopses via the Claude API. The viewer (scripts.app)
    # never imports this; it stays fully offline regardless.
    "anthropic>=0.40",
]
extraction = [
    # Layered document → searchable plaintext extraction
    # (`scripts.extraction`). Without this extra, the cascade still
    # works at tier 0 (stdlib + pypdf + the existing ocrmypdf system
    # binary), so a base install matches the prior `scripts.ingest`
    # behavior. Installing this extra unlocks tier-1+ fallbacks for
    # bezier-glyph PDFs, JS-rendered HTML, and image-only inputs.
    #
    # Provider recommendation order (see README "Extraction"):
    #   1. tesseract  (default, local, no GPU, no network)
    #   2. olmocr     (local 7B VLM, GPU recommended) — install via
    #                 the `extraction-vlm` extra below.
    #   3. cloud VLMs (Claude / OpenAI / generic HTTP) — last resort,
    #                 per-case opt-in required.
    "docling>=2",
    "trafilatura>=1.10",
    "playwright>=1.40",
    "pdf2image>=1.17",
    "pytesseract>=0.3.10",
    "Pillow>=10",
]
extraction-vlm = [
    # Local 7B VLM fallback for severely garbled PDF pages. Stays
    # entirely on the user's machine; recommended over cloud VLMs
    # whenever evidence is sensitive (medical, financial, etc.).
    # GPU recommended; CPU-only inference is slow but functional.
    "olmocr>=0.1",
]
extraction-cloud-openai = [
    # OpenAI vision (gpt-4o family) provider for tier-2 PDF VLM.
    # Sends raw page images to the OpenAI API; per-case consent
    # required (see scripts/extraction/consent.py).
    "openai>=1.30",
]

[project.urls]
Homepage = "https://github.com/EvanOchsner/personal-advocacy-toolkit"
Issues = "https://github.com/EvanOchsner/personal-advocacy-toolkit/issues"

[tool.setuptools.packages.find]
where = ["scripts"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint.per-file-ignores]
# Tests use `pytest.importorskip(...)` before real imports (E402), bind
# call results to `result` for readability even when only side-effects
# are asserted (F841), and use short loop variables like `l` in
# comprehensions over label lists (E741).
"tests/**" = ["E402", "F841", "E741"]

[tool.pytest.ini_options]
testpaths = ["tests"]
